--- a/core/package.yaml
+++ b/core/package.yaml
@@ -16,6 +16,7 @@ library:
     - template-haskell
     - text
     - unordered-containers
+    - vector
     - wl-pprint-text
 
 tests: