Results:

Before:
```
benchmarking clique/5  
time                 89.89 μs   (89.76 μs .. 90.07 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 90.13 μs   (90.03 μs .. 90.30 μs)
std dev              438.5 ns   (338.3 ns .. 608.8 ns)
                       
benchmarking clique/10 
time                 328.8 μs   (328.3 μs .. 329.6 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 329.2 μs   (328.6 μs .. 330.7 μs)
std dev              3.182 μs   (1.999 μs .. 5.713 μs)
                       
benchmarking clique/25 
time                 2.224 ms   (2.215 ms .. 2.231 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 2.241 ms   (2.229 ms .. 2.279 ms)
std dev              65.24 μs   (17.66 μs .. 133.1 μs)
variance introduced by outliers: 15% (moderately inflated)
                       
benchmarking clique/40 
time                 5.599 ms   (5.579 ms .. 5.617 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 5.589 ms   (5.577 ms .. 5.603 ms)
std dev              38.74 μs   (29.50 μs .. 57.13 μs)
                       
benchmarking line/5    
time                 91.56 μs   (89.82 μs .. 93.90 μs)
                     0.995 R²   (0.990 R² .. 0.999 R²)
mean                 90.36 μs   (89.26 μs .. 92.76 μs)
std dev              5.105 μs   (2.755 μs .. 9.682 μs)
variance introduced by outliers: 59% (severely inflated)
                       
benchmarking line/10   
time                 362.8 μs   (359.2 μs .. 365.2 μs)
                     0.997 R²   (0.993 R² .. 1.000 R²)
mean                 386.3 μs   (366.8 μs .. 444.3 μs)
std dev              91.96 μs   (4.856 μs .. 158.9 μs)
variance introduced by outliers: 95% (severely inflated)
                       
benchmarking line/25   
time                 3.104 ms   (3.098 ms .. 3.109 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 3.109 ms   (3.104 ms .. 3.114 ms)
std dev              16.57 μs   (13.72 μs .. 21.20 μs)
                       
benchmarking line/40   
time                 10.56 ms   (10.44 ms .. 10.67 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 10.65 ms   (10.58 ms .. 10.75 ms)
std dev              229.4 μs   (165.6 μs .. 359.7 μs)
                       
benchmarking loop/50   
time                 475.8 μs   (473.2 μs .. 478.6 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 474.5 μs   (473.3 μs .. 475.8 μs)
std dev              4.364 μs   (3.545 μs .. 5.817 μs)
                       
benchmarking loop/100  
time                 1.585 ms   (1.412 ms .. 1.841 ms)
                     0.931 R²   (0.895 R² .. 0.999 R²)
mean                 1.446 ms   (1.418 ms .. 1.569 ms)
std dev              156.6 μs   (40.60 μs .. 343.4 μs)
variance introduced by outliers: 73% (severely inflated)
                       
benchmarking loop/500  
time                 14.73 ms   (-3.266 ms .. 22.14 ms)
                     0.150 R²   (0.002 R² .. 0.951 R²)
mean                 34.54 ms   (25.09 ms .. 57.33 ms)
std dev              27.25 ms   (4.433 ms .. 47.12 ms)
variance introduced by outliers: 94% (severely inflated)
                       
benchmarking loop/1000 
time                 84.30 ms   (80.52 ms .. 88.32 ms)
                     0.996 R²   (0.988 R² .. 1.000 R²)
mean                 86.07 ms   (84.73 ms .. 88.63 ms)
std dev              2.905 ms   (1.467 ms .. 4.789 ms)
                       
benchmarking tight/50  
time                 208.7 μs   (205.2 μs .. 214.5 μs)
                     0.994 R²   (0.991 R² .. 0.998 R²)
mean                 210.6 μs   (206.5 μs .. 215.5 μs)
std dev              14.86 μs   (11.59 μs .. 18.62 μs)
variance introduced by outliers: 65% (severely inflated)
                       
benchmarking tight/100 
time                 813.5 μs   (780.8 μs .. 856.0 μs)
                     0.985 R²   (0.977 R² .. 0.993 R²)
mean                 788.3 μs   (772.2 μs .. 814.0 μs)
std dev              64.77 μs   (43.08 μs .. 91.59 μs)
variance introduced by outliers: 65% (severely inflated)
                       
benchmarking tight/500 
time                 17.00 ms   (16.47 ms .. 17.73 ms)
                     0.993 R²   (0.985 R² .. 0.999 R²)
mean                 17.30 ms   (17.08 ms .. 17.67 ms)
std dev              658.7 μs   (473.9 μs .. 1.002 ms)
variance introduced by outliers: 12% (moderately inflated)
                       
benchmarking tight/1000
time                 71.24 ms   (69.86 ms .. 73.08 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 73.05 ms   (72.16 ms .. 74.04 ms)
std dev              1.676 ms   (1.202 ms .. 2.266 ms)
                       
benchmarking parse/50  
time                 6.199 ms   (6.159 ms .. 6.234 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 6.196 ms   (6.171 ms .. 6.219 ms)
std dev              71.33 μs   (58.81 μs .. 88.38 μs)
                       
benchmarking parse/100 
time                 13.54 ms   (12.88 ms .. 14.37 ms)
                     0.981 R²   (0.959 R² .. 0.995 R²)
mean                 12.87 ms   (12.60 ms .. 13.42 ms)
std dev              947.2 μs   (596.0 μs .. 1.420 ms)
variance introduced by outliers: 37% (moderately inflated)
                       
benchmarking parse/500 
time                 67.27 ms   (63.70 ms .. 70.92 ms)
                     0.994 R²   (0.987 R² .. 0.999 R²)
mean                 72.96 ms   (69.69 ms .. 82.73 ms)
std dev              8.440 ms   (3.331 ms .. 14.13 ms)
variance introduced by outliers: 35% (moderately inflated)
                       
benchmarking parse/1000
time                 138.3 ms   (134.2 ms .. 141.1 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 139.6 ms   (138.5 ms .. 141.6 ms)
std dev              2.102 ms   (1.331 ms .. 2.834 ms)
variance introduced by outliers: 12% (moderately inflated)
```

Current:
```
benchmarking clique/5  
time                 22.76 μs   (22.55 μs .. 22.99 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 22.58 μs   (22.49 μs .. 22.79 μs)
std dev              448.0 ns   (260.5 ns .. 773.6 ns)
variance introduced by outliers: 17% (moderately inflated)
                       
benchmarking clique/10 
time                 51.90 μs   (51.50 μs .. 52.23 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 51.05 μs   (50.74 μs .. 51.39 μs)
std dev              1.097 μs   (919.7 ns .. 1.388 μs)
variance introduced by outliers: 18% (moderately inflated)
                       
benchmarking clique/25 
time                 318.9 μs   (309.9 μs .. 327.4 μs)
                     0.997 R²   (0.995 R² .. 0.999 R²)
mean                 311.5 μs   (308.7 μs .. 315.6 μs)
std dev              11.13 μs   (8.276 μs .. 15.62 μs)
variance introduced by outliers: 31% (moderately inflated)
                       
benchmarking clique/40 
time                 1.598 ms   (1.568 ms .. 1.635 ms)
                     0.996 R²   (0.994 R² .. 0.998 R²)
mean                 1.685 ms   (1.661 ms .. 1.710 ms)
std dev              87.16 μs   (74.88 μs .. 103.3 μs)
variance introduced by outliers: 38% (moderately inflated)
                       
benchmarking line/5    
time                 75.18 μs   (75.00 μs .. 75.38 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 75.33 μs   (75.10 μs .. 75.56 μs)
std dev              756.5 ns   (608.7 ns .. 901.4 ns)
                       
benchmarking line/10   
time                 258.7 μs   (257.7 μs .. 260.2 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 260.9 μs   (259.7 μs .. 264.3 μs)
std dev              5.915 μs   (2.681 μs .. 11.12 μs)
variance introduced by outliers: 15% (moderately inflated)
                       
benchmarking line/25   
time                 1.781 ms   (1.772 ms .. 1.790 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 1.785 ms   (1.778 ms .. 1.794 ms)
std dev              25.47 μs   (18.80 μs .. 34.99 μs)
                       
benchmarking line/40   
time                 5.575 ms   (5.550 ms .. 5.598 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 5.585 ms   (5.570 ms .. 5.604 ms)
std dev              50.10 μs   (39.70 μs .. 69.93 μs)
                       
benchmarking loop/50   
time                 462.3 μs   (459.4 μs .. 466.5 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 463.6 μs   (462.3 μs .. 465.4 μs)
std dev              5.505 μs   (4.182 μs .. 7.378 μs)
                       
benchmarking loop/100  
time                 1.372 ms   (1.366 ms .. 1.377 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.379 ms   (1.374 ms .. 1.390 ms)
std dev              26.42 μs   (14.21 μs .. 45.06 μs)
                       
benchmarking loop/500  
time                 21.61 ms   (21.04 ms .. 22.47 ms)
                     0.997 R²   (0.993 R² .. 1.000 R²)
mean                 21.20 ms   (21.05 ms .. 21.53 ms)
std dev              508.2 μs   (173.8 μs .. 859.7 μs)
                       
benchmarking loop/1000 
time                 84.97 ms   (82.90 ms .. 87.82 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 83.22 ms   (82.79 ms .. 84.50 ms)
std dev              1.183 ms   (265.8 μs .. 1.990 ms)
                       
benchmarking tight/50  
time                 202.5 μs   (201.9 μs .. 203.0 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 203.1 μs   (202.5 μs .. 204.7 μs)
std dev              3.099 μs   (1.561 μs .. 5.743 μs)
                       
benchmarking tight/100 
time                 738.7 μs   (716.0 μs .. 772.8 μs)
                     0.995 R²   (0.989 R² .. 1.000 R²)
mean                 724.0 μs   (720.2 μs .. 736.4 μs)
std dev              21.84 μs   (7.696 μs .. 43.74 μs)
variance introduced by outliers: 21% (moderately inflated)
                       
benchmarking tight/500 
time                 18.29 ms   (17.24 ms .. 19.43 ms)
                     0.988 R²   (0.979 R² .. 0.998 R²)
mean                 17.26 ms   (16.97 ms .. 17.72 ms)
std dev              870.4 μs   (530.6 μs .. 1.214 ms)
variance introduced by outliers: 20% (moderately inflated)
                       
benchmarking tight/1000 
time                 71.10 ms   (70.25 ms .. 71.69 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 72.79 ms   (71.89 ms .. 75.32 ms)
std dev              2.373 ms   (768.4 μs .. 3.978 ms)
                       
benchmarking parse/50  
time                 6.291 ms   (6.218 ms .. 6.381 ms)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 6.246 ms   (6.202 ms .. 6.343 ms)
std dev              184.5 μs   (112.1 μs .. 311.8 μs)
variance introduced by outliers: 13% (moderately inflated)
                       
benchmarking parse/100 
time                 12.46 ms   (12.35 ms .. 12.60 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 12.42 ms   (12.34 ms .. 12.52 ms)
std dev              237.0 μs   (194.3 μs .. 309.4 μs)
                       
benchmarking parse/500 
time                 72.49 ms   (68.44 ms .. 75.16 ms)
                     0.998 R²   (0.996 R² .. 1.000 R²)
mean                 69.12 ms   (68.20 ms .. 70.41 ms)
std dev              1.934 ms   (1.031 ms .. 2.605 ms)
                       
benchmarking parse/1000
time                 139.2 ms   (137.2 ms .. 141.1 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 137.8 ms   (137.1 ms .. 138.8 ms)
std dev              1.261 ms   (652.4 μs .. 1.791 ms)
variance introduced by outliers: 11% (moderately inflated)                  
```