@laurenthuberdeau 2022-03-02
Benchmarks with inlining:
```
Benchmark       bench-inline-new  bench-optimizations-with-inline
clique/10       1.387e-4          0.303e-4 -78.18%               
clique/25       1.045e-3          0.183e-3 -82.46%               
clique/40       2.687e-3          0.803e-3 -70.12%               
clique/5        0.411e-4          0.119e-4 -70.94%               
line/10         1.528e-4          1.186e-4 -22.35%               
line/25         1.037e-3          0.631e-3 -39.10%               
line/40         2.973e-3          1.646e-3 -44.63%               
line/5          0.419e-4          0.390e-4  -6.93%               
loop/100        0.550e-3          0.540e-3  -1.88%               
loop/1000       1.968e-2          1.827e-2  -7.14%               
loop/50         2.178e-4          2.155e-4  -1.01%               
loop/500        0.614e-2          0.575e-2  -6.37%               
parse/100       1.292e-2          1.274e-2  -1.38%               
parse/1000      1.378e-1          1.361e-1  -1.22%               
parse/50        0.627e-2          0.616e-2  -1.85%               
parse/500       0.686e-1          0.676e-1  -1.48%               
tight/100       1.647e-4          1.464e-4 -11.10%               
tight/1000      1.122e-2          1.008e-2 -10.20%               
tight/50        0.565e-4          0.529e-4  -6.50%               
tight/500       3.024e-3          2.545e-3 -15.84%              
Geometric mean  1.420e-3          0.961e-3 -32.34%   
```

[bench-inline-new.csv](https://github.com/Simspace/avaleryar/files/8171368/bench-inline-new.csv)
[bench-optimizations-with-inline.csv](https://github.com/Simspace/avaleryar/files/8171369/bench-optimizations-with-inline.csv)